Use thread extension to retrieve recipient
